+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.7.5
+
+## Version 1.7.5 （2022/09/26）
+
+发布功能
+
+1. Kafka、Redis、Elasticsearch、zookeeper添加用户名、密码、证书等验证
+2. Redis新增、编辑后刷新Keys列表
+
 # 1.7.4
 
 ## Version 1.7.4 （2022/09/25）